+++ PliP/src/com/plip/persistence/dao/interfaces/StatusDao.java
@@ -6,10 +6,10 @@ public interface StatusDao {
 	
 	public Long addStatus(Status status);
 	
-	public Status getStatus(int idStatus);
+	public Status getStatus(long idStatus);
 	
 	public void updateStatus(Status status);
 	
-	public void deleteStatus(Integer statusId);
+	public void deleteStatus(long statusId);
 
 }
